others_unet/Swin-Unet-main/requirements.txt
torch
torchvision
numpy
tqdm
tensorboard
tensorboardX
ml-collections
medpy
SimpleITK
scipy
h5py
timm
einops
yacs
pillow
